editor: Split the breadcrumb element into a module directory
The single file had grown past three thousand lines. Layout planning is pure
width arithmetic, the symbol half and the directory half share nothing but the
segment type, and the bar itself is what ties them together, so those are the
seams. A pure move: no behaviour, no names, and no comments changed.
